18.7.20

+++ src/components/page/NetworkManage.vue
@@ -8,13 +8,16 @@
         <div class="container">
             <!--输入关键词搜索部分-->
             <div class="handle-box">
-                <el-select v-model="select_hasPublic" placeholder="选择网络性质" class="handle-select mr10">
+                <el-select v-model="select_hasPublic" placeholder="选择网络性质" class="handle-select mr10" >
                     <el-option key="1" label="公共网络" value="true"></el-option>
                     <el-option key="2" label="非公共网络" value="false"></el-option>
                 </el-select>
                 <el-button type="primary" icon="el-icon-search" @click="search">搜索</el-button>
-                <el-button type="primary"  icon="el-icon-refresh" style="float: right" @click="syncNetwork">同步</el-button>
-                <el-button type="primary" icon="el-icon-circle-plus-outline" style="float: right" @click="handleCreate">创建</el-button>
+                <el-button type="primary" icon="el-icon-refresh" style="float: right" @click="syncNetwork">同步
+                </el-button>
+                <el-button type="primary" icon="el-icon-circle-plus-outline" style="float: right" @click="handleCreate">
+                    创建
+                </el-button>
             </div>
 
             <!--用户信息展示部分-->
@@ -22,17 +25,19 @@
                 :data="netWorkInfo"
                 tooltip-effect="dark"
                 style="width: 100%">
-                <el-table-column prop="name" label="网络名称" >
+                <el-table-column prop="name" label="网络名称">
                 </el-table-column>
-                <el-table-column prop="scope" label="覆盖范围" >
+                <el-table-column prop="scope" label="覆盖范围">
                 </el-table-column>
-                <el-table-column prop="hasIpv6" label="是否开启ipv6" >
+                <el-table-column prop="hasIpv6" label="是否开启ipv6">
                 </el-table-column>
-                <el-table-column prop="hasPublic" label="网络性质" >
+                <el-table-column prop="hasPublic" label="网络性质">
                 </el-table-column>
-                <el-table-column label="操作" >
+                <el-table-column label="操作">
                     <template slot-scope="scope">
-                        <p style="float: left;margin-left: 5px;cursor: pointer"><i class="el-icon-delete" @click="handleDelete(scope.$index, scope.row)"></i></p>
+                        <p style="float: left;margin-left: 5px;cursor: pointer"><i class="el-icon-delete"
+                                                                                   @click="handleDelete(scope.$index, scope.row)"></i>
+                        </p>
                     </template>
                 </el-table-column>
             </el-table>
@@ -50,9 +55,9 @@
 
             <!--创建网络信息的模态框-->
             <div class="networkForm">
-                <el-dialog title="创建网络" :visible.sync="networkFormVisible" >
-                    <el-form :model="networkForm" >
-                        <el-form-item label="网络名称" :label-width="formLabelWidth"  :rules="{
+                <el-dialog title="创建网络" :visible.sync="networkFormVisible">
+                    <el-form :model="networkForm">
+                        <el-form-item label="网络名称" :label-width="formLabelWidth" :rules="{
                              required: true, message: '网络名称不能为空', trigger: 'blur' }">
                             <el-input v-model="networkForm.name" auto-complete="off"></el-input>
                         </el-form-item>
@@ -60,11 +65,32 @@
                              required: true, message: '网络驱动不能为空', trigger: 'blur' }">
                             <el-input v-model="networkForm.driver" auto-complete="off"></el-input>
                         </el-form-item>
-                        <el-form-item label="网络标签" :label-width="formLabelWidth">
-                            <el-input v-model="networkForm.labels" auto-complete="off"></el-input>
+                        <el-form-item label="网络标签"  :label-width="formLabelWidth">
+                            <el-button @click="add()">添加</el-button>
                         </el-form-item>
-                        <el-form-item label="是否开启ipv6" :label-width="formLabelWidth">
-                            <el-input v-model="networkForm.hasIpv6" auto-complete="off"></el-input>
+                        <el-form-item
+                            v-for="(labelsItem, index) in networkForm.labels"
+                            :label="'标签' + index"
+                            :key="labelsItem.key"
+                            :prop="'labels.' + index + '.value'"
+                            :label-width="formLabelWidth"
+                        >
+                            <el-col :span="6">
+                                <div>
+                                    <span style="float: left">key：</span>
+                                    <el-input v-model="labelsItem.aa" style="width:150px;" >
+                                    </el-input>
+                                </div>
+                            </el-col>
+                            <el-col :span="6">
+                                <div>
+                                    <span style="float: left">value：</span>
+                                    <el-input v-model="labelsItem.bb" style="width:150px;">
+                                    </el-input>
+                                </div>
+                            </el-col>
+
+                            <el-col :span="3" > <el-button @click.prevent="del(labelsItem)" >删除</el-button></el-col>
                         </el-form-item>
                     </el-form>
                     <div slot="footer" class="dialog-footer">
@@ -80,63 +106,63 @@
 <script>
     export default {
         name: "NetworkManage",
-        data(){
-            return{
+        data() {
+            return {
                 // 搜索网络输入的名称
-                select_hasPublic:'',
+                select_hasPublic: '',
                 // 网络列表信息
-                netWorkInfo:[{
-                    "id": "ae736f9a7b14dfcc8fe97bd2f57858078ec827d978eb5569863184b1a5ddcf9e",
-                    "name": "none",
-                    "scope": "local",
-                    "driver": "null",
-                    "hasInternal": false,
-                    "hasIpv6": false,
-                    "labels": "{}",
-                    "hasPublic": true,
-                    "userId": null,
-                    "createDate": "2018-07-14 16:53:51",
-                    "updateDate": null
-                }],
+                netWorkInfo: [],
                 // 分页属性
-                currentPage:1,
-                totalCount:0,
+                currentPage: 1,
+                totalCount: 0,
                 // 创建网络发送的信息
-                networkForm:{
-                name: '',
-                driver: '',
-                labels: '',
-                hasIpv6: ''
+                networkForm: {
+                    name: '',
+                    driver: '',
+                    hasIpv6: '',
+                    labels:[{aa:'',bb:''}]
                 },
-                networkFormVisible:false,
-                formLabelWidth:'120px'
+                networkFormVisible: false,
+                formLabelWidth: '120px'
             }
         },
-        methods:{
+        methods: {
+            //增加标签
+            add(){
+
+                this.networkForm.labels.push({aa:'',bb:''});
+            },
+            //删除标签
+            del(item){
+
+                let index = this.networkForm.labels.indexOf(item)
+                if (index !== -1) {
+                    this.networkForm.labels.splice(index, 1)
+                }
+            },
+            formatNetworkData: function (data) {
+                for (let i = 0; i < data.length; i++) {
+                    if (data[i].hasPublic) {
+                        data[i].hasPublic = '公共网络'
+                    } else {
+                        data[i].hasPublic = '非公共网络'
+                    }
+                    if (data[i].hasIpv6) {
+                        data[i].hasIpv6 = '是'
+                    } else {
+                        data[i].hasIpv6 = '否'
+                    }
+                }
+                return data;
+            },
             // 获取网络列表信息
-            getNetworkInfo:function(){
+            getNetworkInfo: function () {
                 this.$axios.get('/network/listAll' + "?current=" + this.currentPage + "&size=5")
-                    .then(response=>{
-                        if(response.data.code == 0){
-                            this.$message.success({
-                                message: "获取网络信息成功！",
-                                showClose: true
-                            })
-                            this.netWorkInfo = response.data.data.records;
-                            for(var i=0; i< this.netWorkInfo.length; i++){
-                                if(this.netWorkInfo[i].hasPublic){
-                                    this.netWorkInfo[i].hasPublic = '公共网络'
-                                }else {
-                                    this.netWorkInfo[i].hasPublic = '非公共网络'
-                                }
-                                if(this.netWorkInfo[i].hasIpv6){
-                                    this.netWorkInfo[i].hasIpv6 = '是'
-                                }else {
-                                    this.netWorkInfo[i].hasIpv6 = '否'
-                                }
-                            }
+                    .then(response => {
+                        if (response.data.code === 0) {
+                            this.netWorkInfo = this.formatNetworkData(response.data.data.records);
                             this.totalCount = response.data.data.total;
-                        }else {
+                        } else {
                             this.$message.error({
                                 message: "获取网络信息失败！",
                                 showClose: true
@@ -148,30 +174,13 @@
                     })
             },
             // 搜索网络名称
-            search:function f() {
+            search: function f() {
                 this.$axios.get('/network/listAll' + "?hasPublic=" + this.select_hasPublic + "&current=" + this.currentPage + "&size=5")
-                    .then(response=>{
-                        // console.log(response)
-                        if(response.data.code == 0){
-                            this.$message.success({
-                                message: "搜索网络信息成功！",
-                                showClose: true
-                            })
-                            this.netWorkInfo = response.data.data.records;
-                            for(var i=0; i< this.netWorkInfo.length; i++){
-                                if(this.netWorkInfo[i].hasPublic){
-                                    this.netWorkInfo[i].hasPublic = '公共网络'
-                                }else {
-                                    this.netWorkInfo[i].hasPublic = '非公共网络'
-                                }
-                                if(this.netWorkInfo[i].hasIpv6){
-                                    this.netWorkInfo[i].hasIpv6 = '是'
-                                }else {
-                                    this.netWorkInfo[i].hasIpv6 = '否'
-                                }
-                            }
+                    .then(response => {
+                        if (response.data.code === 0) {
+                            this.netWorkInfo = this.formatNetworkData(response.data.data.records);
                             this.totalCount = response.data.data.total;
-                        }else {
+                        } else {
                             this.$message.error({
                                 message: "获取网络信息失败！",
                                 showClose: true
@@ -183,15 +192,15 @@
                     })
             },
             // 同步网络信息
-            syncNetwork:function () {
+            syncNetwork: function () {
                 this.$axios.get('/network/sync')
-                    .then(response=>{
-                        if(response.data.code == 0){
+                    .then(response => {
+                        if (response.data.code === 0) {
                             this.$message.success({
                                 message: "同步网络信息成功！",
                                 showClose: true
                             })
-                        }else {
+                        } else {
                             this.$message.error({
                                 message: "同步网络信息失败！",
                                 showClose: true
@@ -203,27 +212,40 @@
                     })
             },
             // 点击创建按钮
-            handleCreate:function(){
+            handleCreate: function () {
                 this.networkFormVisible = true;
             },
             // 创建网络信息
-            createNetwork:function(){
+            createNetwork: function () {
                 this.networkFormVisible = false;
-                this.$axios.post('/network/public/create',{
+
+                this.networkForm.labels.forEach((item,index)=>{
+
+                    if (item.aa ==="" ||item.bb === "") {
+
+                        this.$message.warning("标签"+index+"未填写完成");
+                        this.networkFormVisible = true;
+                        return ;
+                    }
+                });
+                let labelsMap = {};
+                this.networkForm.labels.forEach((item, index) => {
+                    let key = item.aa.toString();
+                    labelsMap[key] = item.bb
+                });
+                this.$axios.post('/network/public/create', {
                     name: this.networkForm.name,
-                    driver:this.networkForm.driver,
-                    labels: this.networkForm.labels,
-                    hasIpv6: this.networkForm.hasIpv6
+                    driver: this.networkForm.driver,
+                    labelsStr: JSON.stringify(labelsMap)
                 })
-                    .then(response=>{
-                        console.log(response)
-                        if(response.data.code == 0){
+                    .then(response => {
+                        if (response.data.code === 0) {
                             this.$message.success({
                                 message: "创建网络信息成功！",
                                 showClose: true
-                            })
+                            });
                             this.getNetworkInfo();
-                        }else {
+                        } else {
                             this.$message.error({
                                 message: "创建网络信息失败！",
                                 showClose: true
@@ -235,44 +257,42 @@
                     })
             },
             // 删除网络信息
-            handleDelete:function(index, row){
-                this.$axios.delete('/network/delete/' + row.id)
-                    .then(response=>{
-                        if(response.data.code == 0){
-                            this.$message.success({
-                                message: "删除网络信息成功！",
-                                showClose: true
-                            })
-                        }else{
-                            this.$message.error({
-                                message: "删除网络信息失败！",
-                                showClose: true
-                            })
-                        }
-                    })
-                    .catch(function (err) {
-                        console.log(err)
-                    })
+            handleDelete: function (index, row) {
+                this.$confirm('确定要删除该网络?', '提示', {
+                    confirmButtonText: '确定',
+                    cancelButtonText: '取消',
+                    type: 'warning'
+                }).then(() => {
+                    this.$axios.delete('/network/delete/' + row.id)
+                        .then(response => {
+                            if (response.data.code === 0) {
+                                this.$message.success({
+                                    message: "删除网络信息成功！",
+                                    showClose: true
+                                })
+                            } else {
+                                this.$message.error({
+                                    message: "删除网络信息失败！",
+                                    showClose: true
+                                })
+                            }
+                        })
+                        .catch(function (err) {
+                            console.log(err)
+                        })
+                }).catch(() => {
+
+                });
+
+
             },
             // 分页
-            handleCurrentChange:function (val) {
+            handleCurrentChange: function (val) {
                 this.$axios.get('/network/listAll' + "?current=" + val + "&size=5")
-                    .then(response=>{
-                        if(response.data.code == 0){
-                            this.netWorkInfo = response.data.data.records;
-                            for(var i=0; i< this.netWorkInfo.length; i++){
-                                if(this.netWorkInfo[i].hasPublic){
-                                    this.netWorkInfo[i].hasPublic = '公共网络'
-                                }else {
-                                    this.netWorkInfo[i].hasPublic = '非公共网络'
-                                }
-                                if(this.netWorkInfo[i].hasIpv6){
-                                    this.netWorkInfo[i].hasIpv6 = '是'
-                                }else {
-                                    this.netWorkInfo[i].hasIpv6 = '否'
-                                }
-                            }
-                        }else {
+                    .then(response => {
+                        if (response.data.code === 0) {
+                            this.netWorkInfo = this.formatNetworkData(response.data.data.records);
+                        } else {
                             this.$message.error({
                                 message: "获取网络信息失败！",
                                 showClose: true
@@ -284,7 +304,7 @@
                     })
             }
         },
-        created(){
+        created() {
             this.getNetworkInfo();
         }
     }
@@ -296,7 +316,7 @@
     }
 
     .handle-select {
-        width: 100px;
+        width: 150px;
     }
 
     .handle-input {
